Crate buf_read_ext

Structs

AsyncStreamUntilToken
Future associated with async_stream_until_token(). Refer to that functions documentation

AsyncStreamUntilTokenOutput
A completed AsyncStreamUntilToken future yields this value, which returns the streams, as well as the number of bytes streamed and whether or not the token was found.

Traits

BufReadExt
Extends any type that implements BufRead with a stream_until_token() function.

Functions

async_stream_until_token
Returns a future that streams all bytes to out until the token delimiter or EOF is reached.